requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
authlib==1.6.0
bleach==6.2.0
bokeh==3.7.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.3.0
fastmcp==2.10.2
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
linkify-it-py==2.0.3
markdown==3.8.2
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.10.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.46.0
numpy==2.3.1
openai==1.93.0
openai-agents==0.1.0
openapi-pydantic==0.5.1
packaging==25.0
pandas==2.3.1
panel==1.7.4
param==2.2.1
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyviz-comms==3.0.6
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
uvicorn==0.35.0
webencodings==0.5.1
xyzservices==2025.4.0
